vllm integration and general facts test files
This commit implements major changes to how the LLM is leveraged.  It integrates a back end vllm server that is capable of serving multiple threads and batches the requests.  Major token/sec increase as batch size scales.   VRAM requirements have increased as a result.  A future commit will make this option or fall back to ollama. vllm is putting out ~300 tokens/sec across two 3090s on a Mistral-7b fine tune

The prompts file has been updated to work with ollama and also includes a number of improvements to the text in the prompt (more examples).  Code to capture json between ```json <capture> ``` has been added as well as a bunch of other LLM malformed output recovery attempts.